data_gather/challenger_stats/requirements.txt
requests
time
os
python-dotenv
